get rid of angular/vue
