+++ packages/survey-creator-angular/src/angular-ui.ts
@@ -19,6 +19,9 @@ export * from "./side-bar/object-selector.component";
 export * from "./side-bar/property-grid.component";
 
 export * from "./components/component-container.component";
+export * from "./components/icon-item.component";
+export * from "./side-bar/side-bar-launch-card.component";
+export * from "./tabs/designer/designer-surface.component";
 export * from "./components/search.component";
 export * from "./tabs/json/json-editor-textarea.component";
 export * from "./tabs/json/json-editor-ace.component";
