+++ packages/survey-creator-angular/src/tabs/designer/designer.component.html
@@ -6,23 +6,7 @@
   </div>
   <div class="svc-tab-designer" [class]="model.getRootCss()" (click)="model.clickDesigner()">
     <sv-scroll>
-      <div class="svc-tab-designer_content">
-        <ng-container *ngIf="model.showPlaceholder">
-          <div *ngIf="creator.showHeaderInEmptySurvey && creator.showSurveyHeader" class="svc-designer-header">
-            <div *ngIf="survey.renderedHasHeader" [class]="survey.css.header" [survey]="survey" sv-ng-survey-header>
-            </div>
-          </div>
-          <div class="svc-designer__placeholder-container" [attr.data-sv-drop-target-survey-element]="'newGhostPage'">
-            <svc-surface-placeholder [name]="'designer'" [placeholderTitleText]="model.placeholderTitleText"
-              [placeholderDescriptionText]="model.placeholderDescriptionText">
-            </svc-surface-placeholder>
-            <svc-page-wrapper [page]="model.newPage" [creator]="creator" [isGhost]="true" [className]="'svc-designer-placeholder-page'"></svc-page-wrapper>
-          </div>
-        </ng-container>
-        <ng-container *ngIf="!model.showPlaceholder">
-          <svc-designer-survey [model]="model"></svc-designer-survey>
-        </ng-container>
-      </div>
+      <ng-template [component]="{name: model.surfaceComponentName, data: { model: model.surfaceData || model }}"></ng-template>
     </sv-scroll>
   </div>
 </ng-template>
