--- packages/survey-creator-angular/tsconfig.presets.json
@@ -1,13 +1,0 @@
-{
-    "extends": "./tsconfig.json",
-    "compilerOptions": {
-        "paths": {
-            "survey-creator-angular": [
-                "./build"
-            ]
-        }
-    },
-    "include": [
-        "./src/presets/index.ts"
-    ]
-}
