--- packages/survey-creator-angular/ui-preset-editor/ng-package.json
@@ -1,8 +1,0 @@
-{
-  "$schema": "node_modules/ng-packagr/ng-package.schema.json",
-  "dest": "./build",
-  "deleteDestPath": false,
-  "lib": {
-    "entryFile": "src/index.ts"
-  }
-}
