--- packages/survey-creator-angular/ui-preset-editor/src/index.ts
@@ -1,7 +1,0 @@
-export { TabPresetsComponent } from "./presets.component";
-export { PresetsPropertyGridWrapperComponent } from "./presets-property-grid.component";
-export { PresetsIconItemComponent } from "./presets-icon-item.component";
-
-// Export presets module
-import { PresetsModule } from "./presets.module";
-export { PresetsModule };
