+++ packages/survey-creator-vue/example/src/components/Example.vue
@@ -1,7 +1,6 @@
 <script lang="ts" setup>
 import { SurveyCreatorModel } from 'survey-creator-core';
 import { SurveyCreatorComponent } from "survey-creator-vue";
-import "survey-creator-vue/ui-preset-editor";
 import "survey-creator-core/ui-preset-editor/index.css";
 import { UIPresetEditor } from "survey-creator-core/ui-preset-editor";
 const creator = new SurveyCreatorModel({ pageEditMode: "bypage", showLogicTab: true, showJSONEditorTab: true, showTranslationTab: true, showThemeTab: true });
