+++ packages/survey-creator-vue/package.json
@@ -10,13 +10,9 @@
     "serve:example:dev": "npm run dev --prefix ./example",
     "serve:example:prod": "npm run preview --prefix ./example",
     "build:example:prod": "npm run build --prefix ./example",
-    "build": "vite build && npm run build:types && npm run build:presets",
+    "build": "vite build && npm run build:types",
     "build:dev": "vite build --mode development",
-    "build:presets": "vite build --config vite.presets.config.ts && npm run build:types:presets",
-    "build:presets:dev": "vite build --config vite.presets.config.ts --mode development && npm run build:types:presets",
-    "watch:presets": "vite build --config vite.presets.config.ts --watch",
     "build:types": "vue-tsc --project tsconfig.types.json --emitDeclarationOnly",
-    "build:types:presets": "vue-tsc --project tsconfig.types.presets.json --emitDeclarationOnly",
     "type-check": "vue-tsc --project tsconfig.types.json --noEmit",
     "test": "vitest --no-threads tests/test.spec.ts",
     "test:single": "vitest --no-watch --no-threads tests/test.spec.ts",
