--- packages/survey-creator-vue/tsconfig.presets.json
@@ -1,13 +1,0 @@
-{
-    "extends": "./tsconfig.json",
-    "compilerOptions": {
-        "paths": {
-            "survey-creator-core": [
-                "./build"
-            ]
-        }
-    },
-    "include": [
-        "./src/ui-preset-editor/index.ts"
-    ]
-}
